requirements.txt
streamlit
google-generativeai
pillow<12
openai
flask
flask-cors
groq
pillow
python-dotenv